refactor(api,plateform): use shared dto for user-scoring and matching
