+++ packages/dto/src/resources/index.ts
@@ -1,1 +1,3 @@
 export * from "./mission-browse";
+export * from "./mission-match";
+export * from "./user-scoring";
