+++ packages/dto/src/resources/user-scoring.ts
@@ -1,0 +1,9 @@
+export type UserScoringCreateResponse = {
+  id: string;
+};
+
+export type UserScoringUpdateResponse = {
+  user_scoring_id: string;
+  created_count: number;
+  mission_alert_enabled: boolean;
+};
