+++ plateform/app/services/matching.ts
@@ -1,6 +1,7 @@
-import type { MatchResponse } from "~/types/matching";
+import type { MissionMatchResponse } from "@engagement/dto";
+
 import api from "~/services/api";
 
-export async function fetchMatches(userScoringId: string, limit = 5, offset = 0): Promise<MatchResponse> {
-  return api.get<MatchResponse>(`/missions/match?userScoringId=${encodeURIComponent(userScoringId)}&limit=${limit}&offset=${offset}`);
+export async function fetchMatches(userScoringId: string, limit = 5, offset = 0): Promise<MissionMatchResponse> {
+  return api.get<MissionMatchResponse>(`/missions/match?userScoringId=${encodeURIComponent(userScoringId)}&limit=${limit}&offset=${offset}`);
 }
