+++ plateform/app/utils/mission.ts
@@ -1,5 +1,4 @@
-import type { MissionBrowse, MissionDetailCompensation } from "@engagement/dto";
-import type { MatchResultItem } from "~/types/matching";
+import type { MissionBrowse, MissionDetailCompensation, MissionMatchItem } from "@engagement/dto";
 
 const MONTHS = ["janvier", "février", "mars", "avril", "mai", "juin", "juillet", "août", "septembre", "octobre", "novembre", "décembre"];
 
@@ -53,7 +52,7 @@ export function formatMissionType(type: string | null): string {
   return MISSION_TYPE_LABELS[type] ?? "Mission";
 }
 
-export function matchResultToBrowseMission(item: MatchResultItem): MissionBrowse {
+export function matchResultToBrowseMission(item: MissionMatchItem): MissionBrowse {
   return {
     id: item.mission.id,
     title: item.mission.title,
